Move config to module, separate out implementation
Does what it says on the tin. Moves config implementations out to
individual files, guards those around platforms to make stuff more
reasonably split out.
